requirements.txt
apache-airflow>=2.7.1
apache-airflow-providers-amazon>=5.5.0
apache-airflow-providers-postgres>=5.0.0
pandas>=2.0.0
pyarrow>=12.0.0
boto3>=1.26.0
psycopg2-binary>=2.9.9
psycopg2>=2.9.10
dotenv>=0.9.9
requests>=2.32.3

